Can now read multi word game names, and both arrays and string scripts, game menu functional

+++ setup.sh
@@ -9,7 +9,11 @@ display_main_menu() {
     "Select All" "" \
     "Deselect All" "" \
     "Applications" "" \
+    "Package Removal" "Removing unecessary packages." \
     "Games" "" \
+    "Remote Data" "" \
+    "Configs" "" \
+    "Miscellaneous" "" \
     "Setup/Install Selected" "" \
     "Exit" "" 
     # The option selected is printed to stderr. Which has been redirected to stdcout.
@@ -49,7 +53,12 @@ executeSelected() {
     # Execute games commands.
     for (( i=0; i<${#gamesName[@]}; i++ )); do
         if [[ "${gamesSelected[$i]}" == "ON" ]]; then
-            script="$(jq --arg name "${gamesName[$i]}" -r '.[] | select(.name == $name) | .installScript | join("\n")' ./data/games.json)"
+            # If string return string, if array join the array of strings.
+            if [[ $(jq --arg name "${gamesName[$i]}" -r '.[] | select(.name == $name) | .script | type' ./data/games.json) == "array" ]]; then
+                script="$(jq --arg name "${gamesName[$i]}" -r '.[] | select(.name == $name) | .script | join("\n")' ./data/games.json)"
+            else
+                script="$(jq --arg name "${gamesName[$i]}" -r '.[] | select(.name == $name) | .script' ./data/games.json)"
+            fi
             #echo $script
             bash -c "$script"
         fi
@@ -57,15 +66,24 @@ executeSelected() {
 }
 
 # Starting program.
-sudo apt install jq -y # For interpreting json files.
+if ! dpkg -s jq &>/dev/null; then
+    sudo apt install jq -y # For interpreting json files.
+fi
 
 # Read name data from json files, this will be used later to get specific data for each menu. Uses parallel arrays to store whether the corresponding option is enabled. 1st array names, 2nd array whether its enabled, 3rd array scripts, these can be changed when overwrite is enabled/disabled.
+# Application data
+# Removing unecessary packages
+# Game data
 mapfile -t gamesName < <(jq -r '.[].name' ./data/games.json)
-gamesScript=$(jq '.[].installScript' ./data/games.json) # All entries must have a name and script, everything else is optional.
+gamesScript=$(jq '.[].script' ./data/games.json) # All entries must have a name and script, everything else is optional.
 gamesSelected=()
 for ((i=0; i<${#gamesName[@]}; i++)); do
     gamesSelected[i]="OFF"
 done
+# Remote data
+# Configs
+# Miscellaneous - e.g. deleting unecessary folders.
+
 #echo ${gamesName[*]}
 
 while [ "$option" != "Exit" ] 
@@ -80,13 +98,17 @@ do
             ;;
         "Applications")
             ;;
+        "Package Removal")
+            ;;
         "Games")
             # Enter the game menu. return the names of any options that change.
             option="$(display_game_menu 3>&1 1>&2 2>&3)"
             # Split option into individual names.
-            IFS=' ' 
-            read -a options <<< "$option" 
+            #IFS=' ' 
+            #read -a options <<< "$option"
+            eval "options=($option)"
             for opt in "${options[@]}"; do
+                #echo $opt
                 opt=$(echo "$opt" | xargs)
                 index=$(getIndex "$opt")
                 if (( index != -1 )); then
@@ -97,7 +119,13 @@ do
                     fi
                 fi
             done
-            echo "${gamesSelected[@]}"
+            #echo "${gamesSelected[@]}"
+            ;;
+        "Remote Data")
+            ;;
+        "Configs")
+            ;;
+        "Miscellaneous")
             ;;
         "Setup/Install Selected")
             executeSelected
